@@ -12,23 +12,18 @@
   },
   "scripts": {
     "start": "ng serve",
-    "clean": "rm -rf dist",
     "lint": "npm run lint:ts && npm run lint:css",
-    "build": "rm -rf dist/* && npm run build:aot",
+    "build": "ng build --configuration=production",
     "deploy": "rm -rf dist && deploy-to-git",
     "lint:ts": "tslint -p src/tsconfig.json \"src/**/*.ts\"",
-    "lint:css": "stylelint --syntax scss \"src/**/*.css\"",
-    "build:aot": "webpack --config config/webpack-build-aot.config.js",
-    "build:jit": "webpack --config config/webpack-build-jit.config.js",
-    "serve:sjs": "lite-server -c config/lite-server-bs-config.js",
-    "serve:wds": "webpack-dev-server --open --inline --config config/webpack-server-jit.config.js"
+    "lint:css": "stylelint --syntax scss \"src/**/*.css\""
   },
   "config": {
     "deployToGit": {
       "repository": "git@github.com:zefoy/ngx-swiper-wrapper.git",
       "branch": "gh-pages",
       "folder": "dist",
-      "script": "npm run build",
+      "script": "npm run build -- --base-href=ngx-swiper-wrapper --delete-output-path=false",
       "commit": "Publishing $npm_package_version",
       "user": {
         "name": "ZEF Devel",
@@ -37,42 +32,32 @@
     }
   },
   "dependencies": {
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/flex-layout": "7.0.0-beta.23",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "core-js": "^2.6.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/flex-layout": "8.0.0-beta.26",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "core-js": "^3.1.0",
     "ngx-swiper-wrapper": "^7.2.0",
-    "rxjs": "^6.3.0",
-    "swiper": "^4.4.0",
-    "zone.js": "^0.8.0"
+    "rxjs": "^6.5.0",
+    "swiper": "^4.5.0",
+    "zone.js": "^0.9.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.0",
-    "@angular/cli": "^7.2.0",
-    "@angular/compiler-cli": "^7.2.0",
-    "@ngtools/webpack": "^7.2.0",
-    "@types/node": "^10.12.0",
-    "@types/swiper": "^4.2.0",
-    "codelyzer": "^4.5.0",
-    "connect-history-api-fallback": "^1.6.0",
-    "copy-webpack-plugin": "^4.6.0",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@ngtools/webpack": "^8.0.0",
+    "@types/node": "^12.0.0",
+    "@types/swiper": "^4.4.0",
+    "codelyzer": "^5.1.0",
     "deploy-to-git": "^0.1.0",
-    "html-webpack-plugin": "^3.2.0",
-    "lite-server": "^2.4.0",
-    "plugin-typescript": "^8.0.0",
-    "raw-loader": "^1.0.0",
-    "stylelint": "^9.9.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-order": "^2.0.0",
-    "systemjs": "^0.21.0",
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.0",
-    "webpack": "^4.28.0",
-    "webpack-cli": "^3.2.0",
-    "webpack-dev-server": "^3.1.0"
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-order": "^3.0.0",
+    "tslint": "^5.17.0",
+    "typescript": "~3.4.0"
   }
 }
