@@ -35,34 +35,34 @@
     "url": "https://github.com/zefoy/ngx-swiper-wrapper.git"
   },
   "dependencies": {
-    "swiper": "^4.4.0"
+    "swiper": "^4.5.0"
   },
   "devDependencies": {
-    "@angular/cli": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular-devkit/build-angular": "^0.12.0",
-    "@types/node": "^10.12.0",
-    "@types/swiper": "^4.2.0",
-    "codelyzer": "^4.5.0",
-    "core-js": "^2.6.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@types/node": "^12.0.0",
+    "@types/swiper": "^4.4.0",
+    "codelyzer": "^5.1.0",
+    "core-js": "^3.1.0",
     "cpx": "^1.5.0",
-    "rollup": "^1.1.0",
-    "rollup-plugin-node-resolve": "^4.0.0",
+    "rollup": "^1.14.0",
+    "rollup-plugin-node-resolve": "^5.0.0",
     "rollup-plugin-sourcemaps": "^0.4.0",
-    "rxjs": "^6.3.0",
-    "stylelint": "^9.9.0",
-    "stylelint-config-standard": "^18.2.0",
-    "stylelint-order": "^2.0.0",
-    "terser": "^3.14.0",
-    "tsickle": "^0.34.0",
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.0",
+    "rxjs": "^6.5.0",
+    "stylelint": "^10.1.0",
+    "stylelint-config-standard": "^18.3.0",
+    "stylelint-order": "^3.0.0",
+    "terser": "^4.0.0",
+    "tsickle": "^0.35.0",
+    "tslint": "^5.17.0",
+    "typescript": "~3.4.0",
     "watch": "^1.0.0",
-    "zone.js": "^0.8.0"
+    "zone.js": "^0.9.0"
   },
   "peerDependencies": {
     "@angular/common": ">=5.0.0",
