@@ -48,8 +48,8 @@ export class AppComponent {
     hideOnClick: false
   };
 
-  @ViewChild(SwiperComponent) componentRef?: SwiperComponent;
-  @ViewChild(SwiperDirective) directiveRef?: SwiperDirective;
+  @ViewChild(SwiperComponent, { static: false }) componentRef?: SwiperComponent;
+  @ViewChild(SwiperDirective, { static: false }) directiveRef?: SwiperDirective;
 
   constructor() {}
 
